Fix config type

+++ CustomRoundEndConditions/Config.cs
@@ -35,18 +35,18 @@ namespace Mistaken.CustomRoundEndConditions
         /// Gets or sets the percentage of Class D to escape for CI to win.
         /// </summary>
         [Description("Class D Escape percentage for CI to win")]
-        public decimal ClassDEscape { get; set; } = 50;
+        public float ClassDEscape { get; set; } = 50;
 
         /// <summary>
         /// Gets or sets the percentage of Scientists to escape for MTF to win.
         /// </summary>
         [Description("Scientists Escape percentage for MTF to win")]
-        public decimal ScientistsEscape { get; set; } = 50;
+        public float ScientistsEscape { get; set; } = 50;
 
         /// <summary>
         /// Gets or sets the percentage of Scientists to escape for MTF to win when only MTF is alive.
         /// </summary>
         [Description("Scientists Escape percentage for MTF to win when only MTF is alive.")]
-        public decimal ScientistsEscapeOnlyMtfAlive { get; set; } = 50;
+        public float ScientistsEscapeOnlyMtfAlive { get; set; } = 50;
     }
 }
